utils: add method argument to fake_pha
Pass through the setting to sherpa.astro.fake.fake_pha. This is
primarily for testing, but it may be useful for users to see what
the predicted model looks like. It also now matches the fit
call, which accepts a method argument.

Of course, this only works because the counts field of a DataPHA
object can be set to a float, which technically is only valid for
rates, not in all cases.

Some of the tests have had additional commentary added to explain
the expected signal (fixing some places where the existing text
was wrong).