ui: add set_rng/get_rng methods that do nothing
Until now, Sherpa has used what is now the legacy random-number
generation scheme provided by NumPy: see
https://numpy.org/doc/stable/reference/random/legacy.html but note
that this only happened in NumPy 1.17 and Sherpa has been using
NumPy for a lot longer than that.

So, how can we move to the new Generator interface? Well, there's
two parts

- UI access
- lower-level methods

The lower-level methods can send the generator around as an argument,
but this approach doesn't really work well for the UI level, where
it would make sense for the session to carry around the generator
without the user having to worry about it.

In preparation for this, this commit adds get_rng and set_rng
methods that the user can use to change the generator. At this
time they do nothing other than verify the arguments are sensible.

One thought would be for set_rng to accept an integer, which would
then act like numpy.random.set_seed. I can see the advantage, but
I'm not yet sure whether it's worthwhile (it could be passed to
numpy.random.default_rng as a convenience).

The reset_seed fixture used in testing is a canary for the tests
that will need reviewing. It is not yet clear how this fixture
will need to be updated.