Add rng parameter to most random-number-using routines
In most cases we can send in a rng parameter which, if set to None
(the default), will use the legaxy NumPy random API - e.g. call
np.random.xxx directly. If set, then rng is assumed to be a
generator, and used to access the random scheme directly.

Issues are:

a) this is not used by the "higher-up" routines yet
b) it is not used for the sherpa.sim.simulate.LikelihoodRatioTest
   because this needs more work to determine how best to handle
   parallel random-number generation.
c) the main area not addressed here is sherpa.optmethods

A number of routines have been added to sherpa.utils.random to
help in this transition (that is, when the rng parameter may be
None). Thses are intended to be "short-term" features as they
can be replaced with dirct calls to the object.