utils: move parallel code into sherpa.utils.parallel
The sherpa.utils code is rather diverse, so this commit moves out
the "parallel" code into sherpa.utils.parallel, and then imports
it so it's still visible from sherpa.utils (although the idea would
be to change the callin code to use the new location).

There are some small style changes to the code (mainly the use of
f-strings and replace a lambda with an actual function, which
does not need to be senf the data to process as it's already in
scope). There is one change to avoid sending a Lock object around
as we end up not using it.

One interesting part of this is that the multi-processing setup has
been moved out of sherpa/utils/__init__.py. In #975 I suggested
moving this code into a separate module. For this commit I decided
not to do this, to make it easier to trace the behavior.