parallel: add parallel_map_rng
Add a parallel_map_rng routine (currently unused). Improve the
documentation of the existing parallel_map-like code. Simpify
some of the code (e.g. add the new process_tasks routine as it
is used in both versions of run_tasks - the sherpa.utils.parallel
version (generic) and the sherpa.estmethods version (specialized).

As part of this, the sherpa.utils[.parallel]run_tasks routine
has had it's last argument made optional, as it's the same as the
length of the first argument. The argument coukd have been removed
but that would break any code that uses this, so for now we make
this optional with the plan to remove it at some point in the
future (but do not add any deprecation notice at this time).