optmethods: code style changes to ncoresde.py
Some minor code-style chanegs, including

- removing un-needed return lines
- remove __init__ definitions that are just pass
- remove __init__ definitions that just call the superclass
- use a f-string
- for the 'if __name__ ...' section, avoid creating a global
  variable called npar as this is an argument to many
  constructors

There is one TODO that has been added. The pylint tool pointed out
that one constructor was not sending in the seed value it should
be. This could be fixed, but it suggests that there's no code
testing this, so it is not worth fixing at this time (as there may
be other problems with the code).

The 'if __name__ ...' logic should probably be replaced by actual
tests (see #1737), but this is also out-of-scope at this time.
Running the file directly creates many deprecation warnings about
the seed setting, which suggesta that this code does not have much
test coverage in Sherpa.