optmethods: remove mutable argument from OptMethod.fit
This is primarily to clean up the code in __init__.py -
so basically f-string changes - but pylnt pointed out that
the fit method of OptMethod used a mutable argument for
the statskargs argument (which was set to an empty dictionary).
This has been addressed by making the parameter default to None,
which is then converted to {} in the routine.