optmethods: remove Lock call in multiprocessing
To match the recent sherpa.utils.parallel changes, the Lock
argument has been removed from the multi-processing code as it
is unused. There's also an attempt to make the methods - in this
case my_worker - better typed by having more-explicit arguments.
More could be done, but I'm not sure it's worth it here.