Tests: add a low-level test of the Simplex class
When looking at updating the code to use the 1.17 NumPy Generator
class, I thought there was a problem with the seed handling of the
sherpa.optmethods.opt.SimplexBase family of classes. It turned out
to be a false alarm, but this test is one way to check this.

It is very likely that any changes to the RNG will cause the test
to need to be updated - i.e. that the RNG values will be different.