opt: add more support for passing around a RNG
The existing code passes a seed around, and it is used in two
places:

- to control the random numbers from the random module
- sent to C++ code for the MT random code

We do not want to (or, rather, are not going to) change the second
option. For these we just need to ensure we send in a seed (and,
if necessary, have this generated randomly).

The first option can be changed. This means that any code using a
particular seed will no-longer be repeatable, as we are changing
from random.seed to numpy.random.seed (when not sending in an
actual generator).

This is hard to check and verify because there is limited
coverage of this code and it's not obvious what is going on, or
whether the seed/generator is being sent around. There's also
at least two places that looks broken that have been noted but
not changed here

- a seed argument is sent in but it is then replaced by a fixed
  value
- a call is missing a required seed argument