Tests: what happens with keyword-only arguments?
Found when working on #1836, we do not handle routines like

    def do_the_foo(self, foo, morefoo=False, *, manyfoo=None):

which is a "relatively new" construct in Python (compared to
Sherpa). So let's start with a few tests of the functionality
(which involves inspecting the signature of functions/methods)
to show the problem.

Of course, these tested interfaces don't really meed the new
capability: is the morefoo argument labelled a positional or
keyword argument?