utils: support KEYWORD_ONLY parameters
Our existing code to query functions - used for some of the
code introspection we do - does not really understand the
concept of keyword-only parameterd - e.g. for a function
signature like

    def foo(a, b, *, c, d=False):

is c treated as positional or keyword by these routines? I
have chosen positional, since the existing code assumes that
a keyword has a default value, and c does not.

An alternative would be to rework this code introspection, but
let's think about that as a future impprovement.