fit: move call-backs into top-level classes     WIP
The current code works fine with the fork multiprocessig option
but it fails when run with forkserver or spawn because the
pickler used (ForkingPickler) can not identify the local symbols
used. So, we can move these local symbols (these are generally
lambdas used to send context round along with a function to the
optimisation routines) into module-level symbols. The approach
taken is generally to switch from using a local decorator (which
has access to the local context) to a class where we send the
context to the __init__ call and then can use that in the
__call__ call (that is, we treat the returned object as a
function). I did some basic timing tests and it appears that
neither approach has an advantage, but I should warn

- the tests were very basic
- they did not treat all the complexities (e.g. the use of
  a decorator written within the context of another function
  or the cost of sending around objects)
- I can well imagine that it depends on the Python version