fit: minor code cleanup
I had noted a while ago that StatInfoResults was a subclass of
NoNewAttributesAfterInit but did not initialize it during it's
initialization. I have therefore decided to change this to use
the __slots__ technique to mark this as "no new attributes" -
see issue #2008.

There are some typing rules added, and a minor code clean up.