opt: add minimal typing rules for the optimisers  WIP
Two main changes

 1. replace numpy. with np.
 2. add typing rules to record that the optimisers return a
    5-element tuple (and then the follow on cleanup as we trace
    those calls through the fit/optimisation classes).

As part of 2 there is some clean up to create this tuple directly
rather than building it up in parts, and also avoiding decomposing
this return value via list and then recreating it with tuple.

There are also some minor numpy-related cleanups.

There are also some mypy-related changes which may be a wash on
whether the code is improved or not.