optfcts: extract out the 'infinite potential' idea
We have two places where the optimizer checks that the parameter
values are still valid (i.e. within their constraints) and, if
not, return "infinity". So abstract that out into a callable
class. This could have been written as a decorator, but it's
easier (for me) to see this as a class.

I note that by "infinity" I mean "a really large number that we
don't expect to be anywhere close to the statistic values that
are created during fitting", which we take to be the maximum
value of float64 (this limit is used elsewhere in the module,
and is the behavior before this commit so nothing has changed,
I just wanted to note this).