optmethods: add basic typing for the optimization code
We add some basic types to describe the optimization functions:

- OptResults: the return value of the optimizers
  (this moves the local value just added to optfcts to the
   sherpa.utils.types module, along with some documentation)

- StatFunc: a routine you can call with a set of parameters,
  along with possible optional arguments, and get the statistic
  and "per bin" statistic values

- OptFunc: the optimization routine. As with StatFunc there's
  the required arguments and then "un-typed" optional arguments.

The idea is to "be useful" in describing the interfaces, even
if they do not quite cover everything. These are paragmatic
choices - e.g. use of Callable for most but Protocol for OptFunc,
as we try to explain the interface.

The idea is to provide the minimal typing for the optmethods
routines for use by the fit module. However, the modules still
need work to properly type the routines.