optfcts: use a single routine to create the return type
Ideally we'd use some form of structured type for returning the
optimiser data, but that would be a large change at this point
and we have many other changes to do before we can get there.
So, instead, create a little routine that will create the return
value, since this centralizes the code

- to make it easier to change to return whatever we decide
- makes it more obvious how the various fields are set

I am not competely happy with the code, since there are different
ways this could be done with a different level of code duplication.
In particular, can we just move the _get_saofit_msg logic into
the new _optval routine? For now I do not think it's worth spending
too much time reworking this code.