opt: clarify arguments to Opt
Parts of the Opt initialization allowed xmin and xmax to be
None, *but* the code started by calling len(xmin) and then
using that if either xmin or xmax were not set. This
doesn't make sense, so make sure xmin and xmax are set - that
is, it is up to the caller to make sure that the bounds are
set. This allows the code to be cleaned up *and* none of the
tests are changed (so we never tested the old code paths that
have been removed).