optmethods: very minor internal cleanup
One clean up - to return a comparison rather than check the
result of the comparison and then return True or False.

Unfortunately the optmethods code has lots of places where the
code could be cleaned up, but I do not have the time, energy, or
patience to properly do this at this time.