C++: allow grouping logic to be set by Python function
The existing grouping code requires the operation to apply to the
groups to be sent in as a string. This is "fast" but it means that a
user can not apply their own method to the grouping code.

This commit creates a second user-callable routine that lets the user
send in a function - not a string - that takes in an array and returns
a scalar value.

The original code is still used, and the new code is provided as a
means of experimentation.

*** NOTE timing tests may not exactly match this commit as some
         things have changed, but I would be surprised if it would
         make a difference ***

When comparing, for my laptop I get (using two groupnig arrays, of
length 1024 and 8192 channels set up to match a given filtering and
grouping scheme, so does not capture all corner cases but should be
good-enough for initial testing (times are in 10^-6 s):

  Method              1024            8192
  1 "sum"        2.4 +- 0.003      27  +-  0.16
  2 sum         76.6 +- 0.43      1400 +-  1.8
  3 np.sum      74.4 +- 0.22      2990 +- 10.9

where

  method 1 is the all-C++ code (i.e. the original)
  method 2 is the new code called with the Python sum routine
  method 3 is the new code called with the NumPy sum routine

It is interesting that the NumPy version is slower than the Python
version here, but that could perhaps be improved at some loss in
generality in the code.