C++: internal change to the grouping logic
Tweak the new code to create the array object directly from the
underlying data store - after making sure it's contiguous - rather
than copying over the values for each loop. This is slightly
faster but the logic is a bit more complex.

** Times are not guaranteed to match this commit but should be similar **

Using the same set up for timing, we get (times are in 10^-6 s):

  Method              1024            8192
  1 "sum"        2.4 +- 0.003      27  +-  0.16
  2 sum         72.3 +- 0.57      1310 +-  3.4
  3 np.sum      73.4 +- 0.25      3020 +-  4.4

where

  method 1 is the all-C++ code (i.e. the original)
  method 2 is the new code called with the Python sum routine
  method 3 is the new code called with the NumPy sum routine