C++: group function check on NULL rather than a string
The group function to use is either a templated function or
something that can be implemented directly ("_make_groups"). The
original check had used a string comparison, but this is changed
to use a pointer comparison since it feels a more natural check,
in that we only ever set up the func argument if we are going
to use it.