xspec: ensure set_xsabund("file") only after setting abundances
It turns out if you try to set the abundance setting to "file"
before reading in any data then you can crash XSPEC. So add a
simple boolean marker to ensure we can not do this.

An alternative would be to ban "file" from the set_xsabund call,
but maybe people like switching between a file value and another
table (although there are other ways to do this in Sherpa).