TESTS: add linked-parameter checks of normal_sample - issue #2332
Do we get sensible results if we have a linked-parameter in the model
when calling normal_sample? Issue #2332 is only a problem when
correlate=False.

It turns out that linking a parameter changes the order of the thawed
parameters - for valid reasons - which means that it is hard to check
the results (here I have a fit with no linked parameter and one with a
linked parameter), but we can do some validation.

However, we do get a result that I do not understand for the
chi-square (not poisson) statistic: the sampled statistic values are
the same for the two fits (original and linked parameters) which seems
nice at first but I can not explain it. I have added checks of both
these cases here: if they end up failing in the future it may not be a
problem, but it would be useful to know that there has been a change.