CI: bump python versions to 3.11/12/13
Drop Python 3.10 and add 3.13 from the CI runs.

As part of this, drop testing with NumPy version 1. It is not clear
what NumPy versions to use so either force version 2 or use a version
we are currently testing CIAO development code with.

Another change is that the CI runs now just install python rather
than force a cpython variant, as this was causing problems with the
testing of Python 3.13.