WIP try to store the load_xxx commands for serialization
This is primarily to improve handling of PHA2 files but affects a
lot of the code. This is incomplete as I just got to the part of
storing "different" types of load_xxx commands (e.g. "data", "bkg",
"arf", "psf") but haven't really implemented this or tried to
take advantage of it in the serialization code.