TESTS: support HEASARC XSPEC conda builds for XSPEC tests
The existing test suite is based on the CIAO-provided version of
XSPEC, and this has historically laid files out differently to the
HEASARC approach. With HEASARC now providing an XSPEC build via conda,
we can see this difference in two tests.

So, the tests have been updated to allow either layout, as really all
we care about is that the result is "meaningful". It is not worth
trying to use out-of-band data to identify the "expected" layout.