astro: add typing rules to the background module
As well as typing rules there are some small changes related to
using f-strings rather than format statements, and a clarification
of an identifier argument.

The types in this module try to be as general as possible whilst still
being meaningful, which means that ArithmeticModel is chosen as the
"base" class rather than Model (since we need to be able to combine
these models). However, the current UI layer tends to use Model rather
than ArithmeticModel, which leads to the need to use a cast in one
place. Arguably the correct thing to do is change the UI typing
statements, but that is a large job, so for now we use the simpler
cast form.

There are also places where it would be good to say "this has to be a
convolution-style model", but we do not have a good understanding of
how to type this here, so leave these as ArithmeticModel.