background: rework addition of the model components. Fix #2400
Create the get_full_expr routine which will add in the backgound
components to the model expression. There's two modes:

- we can separate out the response and the model
- we have to add the response multiple times because of an array
  of scaling values

determined by whether the list returned in the second argument is
empty or not.

As part of this, add a get_bkg_srcs routine to extract the background
data for a dataset.

Use these to address issue #2400, so that plot_model_components and
get_model_components include the background components. Of course,
there are some cases where this will not work, so warning messages are
generated. The implementation is not ideal, as we need to add logic to
plot_model_components, but it would be nice to somehow make this done
without the need to replicate the top-level code.