TESTS: add a check of plot_model_components for PHA with background
This shows issue #2400 - that the background model is not included
in the plot_model_components display. We want tests with background
data with

- no model
- model with scalar scaling
- model with vector scaling

We could add tests for multiple background components, but assume
that logic has been covered elsewhere; all that needs to be checked
here is that we can access the background data.