ui/utils: minor typing changes for model-related routines  WIP
Add more types for routines that deal with models. There is a
large issue that this uncovers but not dealt with here. Namely,
that we are using Model in places where we should probably be
using a more-accurate type (primarily ArithmeticModel), because:

a) changing to the sub-classes is actually a lot of work,
b) and it's not always obvious what thype to use (e.g.
   CompositeModel or ArithmeticModel).

So for now use Model as the basis.

The ModelWrapper class is an interesting point: can we tell the type
checker that, when called, it returns the same type as the input
class? At the moment we do not try to do this.

pyright 1.1.407 -> 75 errors, 0 warnings, 0 informations