ui/utils: minor improvements to parameter typing rules
This includes priors and samplers.

Note that set_sampler is interesting, because technically we can
send in a class (not an instance of the class), but

a) we don't make the classes easy to find, so it is much easier to
   stick with the string names

   [we do have caes like stat and method where it is easier to get
    at an object to send in, but here it is harder to do]

b) it is actually unclear whether we should send in a class or an
   instance of a class.

   I am 99% sure it is the class, rather than instance, but the
   sherpa.sim code is not well tested.

c) if you do send in a class (or instance) then the code is
   likely to fall over internally, because there is no tests and
   so no-one noticed an issue with this approach.

So I have removed the note about sending in an instance (which also is
likely wrong) and labelled it as just accepting a string.  Once the
sherpa.sim code is properly typed and tested we can look at allowing
more options for the set_sampler argument.

The internals of add_user_pars has been cleaned up to make the type
checker happy, and to make the code a bit cleaner.

pyright 1.1.407 -> 77 errors, 0 warnings, 0 informations