ui/utils: add more typing to file I/O commands
This change requires Python 3.11 or later - to support the last
argument of Concatenate being "...".

These mainly just explicitly add "known" types (e.g. bool and str)
to the calls.

Unfortunately we do not have an easy way to label "file names"
because

- we may just accept str
- it would be nice to allow Path objects
- some calls allow sending in objects like StringIO that have a
  write method.

For now the updates are focussed on those that can be represented
as a string.

pyright 1.1.407 -> 78 errors, 0 warnings, 0 informations