ui/utils: improve typing of plot/contour/image routines
The current plot design makes it hard to know how to type things like
the "data plot", as there's a "plot" and "histogram" variants and
these have different lot hierarchies. There are also issues with the
fact the base Plot/Contour classes take actual data to plot, but most
of the classes require a Data object - via a prepare method - and it
is rather messy. The addition of MultiPlot (which doesn't really fit)
hasn't helped.

So, for now, explicitly type the return types but some of the internal
containers remain as containing Any objects. It will be interesting to
see how this behaves in sherpa/astro/ui/utils.py, where PHA-related
plot classes are added (although, as they are likely sub-classes of
the listed classes it should hopefully be okay).

pyright 1.1.407 -> 46 errors, 0 warnings, 0 informations