ui/utils: add explicit types for the remaining boolean parameters
Should we just remove bool_cast from most places? I worry that this
could break people's code, that have set "clobber='no'" for instance,
for little gain. So, we add a bool_cast_scalar routine and then use
that here. One possibilty would be for it to issue a deprecation
warnig if given a boolean-like-but-not-bool value, but that can be
done at a later time.

So, this is technically a breaking change, as sending in

    clobber=[1, "no"]

is now treated as

    clobber=True

rather than

    clobber=np.ndarray([True, False])

but the original version would have then lead to an error, **and**
it also does not make sense here, so I am okay with this change.

pyright 1.1.407 -> 32 errors, 0 warnings, 0 informations