ui/utils: internal Data type and code changes
This removes the internal _read_data method, which was marked as a
static method, since it is only used in one place so the logic can
just be done there instead (as the method isn't very complex).

A number of routines are now labelled as taking Data objects,
and returning values.

The get_staterror code now errors out if there's no statistical
information (which is only possible if the statistic object has
no calc_staterror routine).

pyright 1.1.407 -> 41 errors, 0 warnings, 0 informations