ui/utils: internal code clean up
Two style changes, and code clean-up change in set_psf and save.

In set_psf the self._psf[id] value is set to be psf in the _set_item
call (which validates that psf is not None), but then we immediately
query self._psf for the id value and allow it to be None. This check
can not succeed, so remove it.

In save, switch from explicit handle management (try/finally) to
using a context manager to deal with the file handle.

pyright 1.1.407 -> 70 errors, 0 warnings, 0 informations