data: internal clean up to the Data hierarchy
The get_filter and get_filter_expr methods are added to the Data class
since it better matches assumptions made in the UI layer (that is,
that all data objects have these methods). They are effectively
virtual (since the implementation is to raise NotImplementedError) but
I have not added abc.abstractmethod decorators because the Data class
does not derive from abc.ABCMeta. Note that the get_filter routine
does not have consistent usage of keyword arguments in the subclasses,
so we mark this by having the routine accept **kwargs.

It is not clear that this is the best approach, since the Data2D class
does not really have the concept of a string-based filter. However,
the sherpa.astro.data.DataIMG class does (via region support), so this
appears to be the simplest approach.

The Data.get_dims method has been made "virtual", in that it will now
raise a NotImplementedError if called, because

a) all the sub-classes over-ride this method
b) all the over-rides return different information

Ideally we could just move all logic into Data.get_dims, but there is
some strange choices made in the Data2D case (which really only comes
into play in the DataIMG class) which means that at present it's best
to keep the code separate.

Some methods which were marked with @overload, to allow the different
calling approaches to be typed, needed to add types to the actual
callable implementation.  So fix this. There is still uncertainty over
how a field like "y: X | None = None" should be typed in the overload
version (when y is None); it appears we do need to say something like
"y: None = None".

A number of methods relating to axis data have seen some minor
tweaks. There is uncertainty in the API over what to do when an axis
is unset: should it error out or return None. For now try to support
the existing code, but there are some cases where we now explicitly
error our rather than causing a run-time error (such as multiplying
None by a value).