ui/utils: start typing the optional id-as-first-argument routines
A number of the methods are labelled as "not being very Pythonic",
since they are allowed to be called like

    method(val [, ...])

where the identifier is taken to be the default value, or

    method(idval, val [, ...])

where the identifier is explicitly included.

It is possible to add types for these routines, but it hasn't been
worth it until recently (i.e. until many other routines have had types
added).

This only allows set_data to be typed, but the helper routine is
written as a generic routine (so it should not need to be re-written
for each "value type").

We could advantage of the @overload decorator to attempt to reduce the
number of combinations of parameters the type checker (and user) has
to consider, but it turns out we do not seem to be able to use this in
the new helper routined (that does the type conversion for us), at
least with pyright 1.1.407.

For now, we do not add @overload versions of the methods since it does
not add much (and in some cases causes problems, such as if we remove
the commented-out overload statements for the separate_id_from_arg
routine).

pyright 1.1.407 -> 26 errors, 0 warnings, 0 informations