ui/utils: support optional id-as-first-argument routines for models
The model routines are hard to type properly because the base class of
Model is often not sufficient, but it's not currently obvious how best
to rework this. So we use Model for now, because it's better than
leaving it as Any.

The routines are also a bit-more varied than the other routines that
have been typed, since the logic for them is not always as
simple (e.g.  the guess routine).

The "pvalue plot" code has seen some typing rules added, but

a) not for the convolution model, since the convolution support
   needs more thinking about how best to type (Model is definitely
   not correct, since it needs to be something that is sent a Model
   and returns a Model)

b) for get_pvalue_plot why do we allow the models to be set to None
   when the code immediately falls over if they are not changed.

pyright 1.1.407 -> 26 errors, 0 warnings, 0 informations