utils: add typing statements
Add typing statements for many, but not all, symbols in sherpa.utils.

This module has a number of cases where we are typing the "intention"
rather than the actual runtime behaviour of the code. This is
particularly noticeable for some of the routines returning a TypeGuard
judgement, since the code is checking for "iterable" but we claim it
is a "sequence", since the result is a better match to how we use the
code (and the current support for abc types).

The numeric code is not changed significantly here as this needs a lot
of care and attention.