sim: add type statements WIP
Add type statements to the top level of the sherpa.sim module.

This shows a logic error in one case that we obviously do not have a
test of. It is not immediately clear what the correct behaviour should
be (probably to grab the name from the sampler) but as we have no test
leave fixing for now.

It's also not clear whether we are sending around the class or an
instance of the class for the Walk code. Needs looknig at.