ui/utils: add types for the ranges for notice/ignore
The types could be a bit-more precise, e.g. to point out that if lo is
a string then hi is None, but leave that out for now.

There's also an internal change to _notice_expr* routines, which used
to be labelled as taking "str | None" when it turned out to only be
sent a string.

pyright 1.1.407 -> 36 errors, 0 warnings, 0 informations