image: add basic typing support
Not everything is typed, in particular the data argument for the
xpaset calls. The backends are also not typed.

There are also some code "refresh" changes - e.g. use of super()
rather than explicitly naming the superclass, and using the
staticmethod decorator rather than as a function, since that seems to
make pyright happier.