TESTS: corner-case handling of ui.utils
There are some un-tested code paths when the data either has no data
at all, or has no dependent axis, and several routines are called with
a scalar value to set a dependent-axis-related value.

Add some tests to check the current behaviour, which are "ugly" errors
resulting in TypeErrors with messages that will not help the user.