add explicit typing for a few table and user model symbols
Indicate that the file/filename attribute of these models can be a
string when set (it would be nice to allow a Path but that is a
much-larger change).

Both table and user models can be sent a "y" array, but the existing
UI code is typed to allow to be None, but that is silly.