image: typing improvements
Add typing statements to the DS9 code. Note that the original DS9 code
supports - or claims to support - different use cases than used by
Sherpa. For now ignore this extra functionality when
appropriate (e.g. xpaset is labelled as accepting a string or a list;
we only use the string version so that is what is used in the type).

As part of this the dummy backend has been changed to use 'return ""'
rather than pass for the two functions labelled as returning a string.