image: rework the class structure
There was a lot of repeated work in the classes derived from Image,
so

- add a "BaseImage" class that sits below the Image class and is for
  all classes that provide a prepare_image method

- as this provides much of the functionality of the sub-classes we
  can remove a lot of code

I orignally had named BaseImage as _Image and was treating it like a
local symbol. However, this broke the class instance diagram, and make
it harder to work out what was going on in the Sherpa documentation,
so the class name was made global.

This is similar to changes I have prototyped with the Plot and Contour
classes, but have yet to implement.