image/DS9: drop the unused dataFunc argument for xpaset
The dataFunc is always None and is not actually used, so remove it.
This is technically a breaking change, as users could have set it, so
removing it would break their code, and there is a possibility for a
change in behaviour (if dataFunc is set but data is not).  However,
Sherpa has never advertised the sherpa.image.DS9 module, so it would
have had to be the most intrepid user who has done this.