TESTS: improve coverage of image/DS9 code
An explicit test of the WCS code is now included. This is potentially
problematic as it may depend on OS (for machine precision) or DS9
version (e.g. on the keys returned and the ordering).

It did show one difference: my local machine had a preference setting
for "wcs sky" to be FK5 but the default is now ICRS (and we use ICRS
for sending in the serialized WCS data to DS9), which caused failures
on the CI machine. The test has been adjusted to now force ICRS when
requesting the data, to avoid this problem.