image: fix minor issues
The low-level errors from xpaget/set did not decode the message,
leaving a trace of the "byte" output (that is, the message included
"b'xxx" rather than "xxx"). This is a hold-over from the Python 3
transition, and is fixed by explicitly decoding the "error" message
(in one case this comes from stdout, thanks to how xpaset works).

The name field of DataImage will get over-written if the data object
has a header with an "OBJECT" keyword. However, this was never
"reset", meaning that it would persist for any DataImage display until
a new data object with an "OBJECT" keyword is found. The prepare_image
call now explicitly resets the name field in this case.  There are
some potential downsides (in particular what happens with the PSF
Kernel structure, but leave that for a later investigation as the code
now ensures the name is "PSF_Kernel" in this case, just in case).