utils: simplify bool_cast and add other boolean type statements
The bool_cast routine has likely long-passed it's purpose - as it was
likely added to support users used to the pre-Python Sherpa as well as
those using Sherpa via the S-Lang bridge. However, it is a bit early
to remove, but we can simplify its implementation as it turns out the
only place that we ever exercise the "convert a sequence of values to
a ndarray of bools" is in a single test. So we can just leave it as
working on scalar values, which greatly reduces the number of type
complaints when used in other parts of the system.

Improve the typing of routines related to booleans. The use of
TypeGuard here is meant to indicate the intent rather than actually
capturing the nuances of the current scheme (since I have not worked
out a nice way for us to say "something we can iterate over", as all
the various options seem to have an issue).

With pyright 1.1.407, the number of typing errors decreases due
to these changes (mainly the bool_cast simplification):

           ui/utils   astro/ui/utils   data   astro/data
  before:    116          160           88       156
  after:     105          131           87       148