utils: generic typing and code improvements
Add types to most of the routines in sherpa.utils. There remains some
numerical code that needs a review, but that is significantly more
effort than the other changes made here, so is being put off until a
later date.

There are a few mior code changes, mainly to try and track the types
of variables a bit better (e.g. avoid having a value that changes its
type too much).