utils: fix up overload statements
The handling of optional parameters, particularly if they are a
boolean and have a default value, is a little tricky to get right.
The changes here were tested out against pyright 1.1.407 and
mypy version 1.18.2.

It looks like we no-longer have to add the "combined" types for the
runtime version of the routine, but leave in.